// _genApi/static-asset/models/AssetInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AssetType } from './AssetType';
export type AssetInfo = {
    id: number;
    ownerId: number;
    assetType: AssetType;
    name: string;
    path: string;
    fileExt?: string;
    createTime: number;
};